[#47] [telepath-js] Moves Telepath mock to `telepath-js` workspace and provides usage instructions
